+++ fpm-alpine/docker-entrypoint.sh
@@ -40,11 +40,14 @@ get_docker_secret() {
     fi
 }
 
+get_docker_secret PMA_USER
 get_docker_secret PMA_PASSWORD
 get_docker_secret MYSQL_ROOT_PASSWORD
 get_docker_secret MYSQL_PASSWORD
 get_docker_secret PMA_HOSTS
 get_docker_secret PMA_HOST
+get_docker_secret PMA_CONTROLHOST
+get_docker_secret PMA_CONTROLUSER
 get_docker_secret PMA_CONTROLPASS
 
 exec "$@"